Dev-201910: module api init sequence, resolve #27

+++ sample/base.common/src/main/java/me/xx2bab/bro/sample/common/api/ISettingsApi.java
@@ -2,9 +2,10 @@ package me.xx2bab.bro.sample.common.api;
 
 import me.xx2bab.bro.sample.common.mine.IMinePresenter;
 
+
 public interface ISettingsApi extends IBaseApi {
 
-    int getPi();
+    double getPi();
 
     IMinePresenter getMineFragment();
